packages/python-packages/apiview-copilot/scripts/infra/requirements.txt
asyncio==3.4.3
azure-ai-resources==1.0.0b9
azure-appconfiguration==1.7.1
azure-cosmos==4.9.0
azure-identity==1.21.0
azure-keyvault-secrets==4.10.0
azure-search-documents==11.5.2
azure-mgmt-appconfiguration==5.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-cognitiveservices==13.7.0
azure-mgmt-cosmosdb==9.7.0
azure-mgmt-keyvault==12.0.0
azure-mgmt-resource>=22.0.0,<23.0.0
azure-mgmt-search==9.1.0
azure-mgmt-web==8.0.0
beautifulsoup4==4.13.3
black==23.1.0
colorama==0.4.6
markdown-it-py==3.0.0
openai==1.67
prompty[azure]==0.1.49
python-dotenv==1.0.1
semantic-kernel==1.27.2
tiktoken==0.9.0
PyYAML==6.0.2
